Added BookingService for GET/ APIs

+++ src/main/java/com/agoda/booking/tracker/dtos/CustomerSummary.java
@@ -2,11 +2,14 @@ package com.agoda.booking.tracker.dtos;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
+
+import java.math.BigDecimal;
+
 @AllArgsConstructor
 @Data
 public class CustomerSummary {
     //add validations
     String customerID;
-    Long numberOfBookings;
-    Double totalPriceInUSD;
+    int numberOfBookings;
+    BigDecimal totalPriceInUSD;
 }
