+++ src/main/java/com/agoda/booking/tracker/dtos/HotelBookingSummary.java
@@ -3,11 +3,14 @@ package com.agoda.booking.tracker.dtos;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 
+import java.math.BigDecimal;
+import java.math.BigInteger;
+
 @AllArgsConstructor
 @Data
 public class HotelBookingSummary {
 
-    private final Integer hotelId;
-    private final Double totalPriceInUSD;
+    private final BigInteger hotelId;
+    private final BigDecimal totalPriceInUSD;
 
 }
